requirements.txt
numpy
matplotlib
Pillow
scipy
urllib3
astropy
